+++ app/config.py
@@ -1,0 +1,7 @@
+import os
+
+class Config:
+    MONGO_URI = "mongodb://mongodb:27017/evansdale_pizza"
+    ORDERS_DIR = "./orders/"
+    DISPLAY_LIMIT = 5
+
