+++ app/db.py
@@ -1,0 +1,6 @@
+from flask_pymongo import PyMongo
+
+mongo = PyMongo()
+
+def init_db(app):
+    mongo.init_app(app)
