+++ app/encoder.py
@@ -1,0 +1,11 @@
+from datetime import datetime
+from bson import ObjectId
+from json import JSONEncoder
+
+class MongoJSONEncoder(JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, ObjectId):
+            return str(obj)
+        if isinstance(obj, datetime):
+            return obj.isoformat()
+        return super().default(obj)
