+++ app/static/css/styles.css
@@ -41,6 +41,17 @@ h1, h2 {
   height: 780px; /* Fits content within a 1080 x 900 window, accounting for header/footer */
 }
 
+#recent-orders {
+  background-color: var(--light-grey-background);
+  padding: var(--padding);
+  border-radius: var(--border-radius);
+  box-shadow: 0 4px 8px var(--box-shadow-color);
+  flex: 1;
+  min-width: 250px;
+  overflow-y: auto; /* Enables scrolling if content overflows */
+  max-height: 700px; /* Constrains height within the section */
+}
+
 #order-summary, #order-form-section, #recent-orders {
   background-color: var(--light-grey-background);
   padding: var(--padding);
@@ -55,17 +66,17 @@ h1, h2 {
 }
 
 #order-form-section {
-  flex-basis: 40%;
+  flex-basis: 45%;
 }
 
 form {
-  width: 100%;
+  width: 98%;
   display: flex;
   flex-direction: column;
 }
 
 input, textarea {
-  width: 100%;
+  width: 90%;
   padding: var(--input-padding);
   margin-bottom: 15px;
   border: 1px solid var(--input-border-color);
@@ -79,7 +90,7 @@ input:focus, textarea:focus {
 }
 
 input[type="number"] {
-  width: 100%;
+  width: 90%;
   max-width: 60px;
   padding: 5px;
   border-radius: 4px;
@@ -104,10 +115,11 @@ button:hover {
 .order {
   border: 1px solid var(--input-border-color);
   border-radius: var(--border-radius);
-  padding: var(--padding);
-  margin-bottom: 15px;
+  padding: 8px; /* Reduces padding for compactness */
+  margin-bottom: 8px; /* Reduces margin between orders */
   background-color: #f9f9f9;
   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  font-size: 0.9em; /* Slightly smaller font to fit more text */
 }
 
 .error {
@@ -126,9 +138,9 @@ button:hover {
   align-items: flex-start;
 }
 
-.pizza-slice-group label {
-  margin-bottom: 5px;
-  font-weight: 500;
+.pizza-slice-group label,
+.pizza-slice-group input {
+  font-size: 0.85em; /* Smaller font for form labels and inputs */
 }
 
 fieldset {
@@ -152,7 +164,9 @@ footer {
 
 /* Styling for the last orders section */
 #lastOrders {
-  margin-top: 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 5px; /* Closer spacing between orders */
 }
 
 .pizza-slice-group label {
@@ -161,5 +175,11 @@ footer {
 }
 
 .summary {
+  margin-bottom: 5px;
+  font-size: 0.9em; /* Smaller font for compact display */
+}
+
+#recent-orders h2 {
   margin-bottom: 10px;
+  font-size: 1.2em;
 }
